PP-6076 Split TransactionResource tests into different classes
To make it easier to identify tests for the search and CSV endpoints,
split these out into separate test classes.
